Refactor plugin source into modular files
